requirements-test.txt
-r requirements.txt
coverage==4.5.3
pytest==4.4.0
pytest-cov==2.6.1
requests-mock==1.5.2
